+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 2.x branch
 ## 2.13 branch
+### 2.13.1
+* added `LoadAllFixturesTrait` and `LogsMethodsTrait` trait for TestSuite;
+* removed `MailHelper::obfuscate()` method;
+* the MIT license has been applied.
+
 ### 2.13.0
 * fixed bug for `isUrl()` detector; 
 * removed all reCAPTCHA classes and libraries. Use instead